Change the way Realm is initialized
