+++ app/src/main/kotlin/com/mgaetan89/showsrage/ShowsRageApplication.kt
@@ -5,10 +5,7 @@ import android.os.Bundle
 import android.support.v7.media.MediaItemStatus
 import android.support.v7.media.MediaSessionStatus
 import android.support.v7.media.RemotePlaybackClient
-import com.mgaetan89.showsrage.helper.Migration
 import com.mgaetan89.showsrage.model.PlayingVideoData
-import io.realm.Realm
-import io.realm.RealmConfiguration
 
 class ShowsRageApplication : Application() {
     var playingVideo: PlayingVideoData? = null
@@ -19,20 +16,6 @@ class ShowsRageApplication : Application() {
         }
 
     fun hasPlayingVideo() = this.playingVideo != null
-
-    override fun onCreate() {
-        super.onCreate()
-
-        Realm.init(this)
-
-        val configuration = RealmConfiguration.Builder().let {
-            it.schemaVersion(Constants.DATABASE_VERSION)
-            it.migration(Migration())
-            it.build()
-        }
-
-        Realm.setDefaultConfiguration(configuration)
-    }
 
     private fun startVideo() {
         val route = this.playingVideo?.route ?: return
